pyproject.toml
[project]
name = "closed-context-retrieval"
version = "0.0.1"
description = ""
authors = [
    {name = "Brian Mendoza", email = "bcmendoza@gmail.com"},
]
dependencies = [
    "sentence-transformers>=2.2.2",
    "redis>=4.6.0",
    "hiredis>=2.2.3",
    "discord-py>=2.3.1",
    "uvloop>=0.17.0",
    "textract>=1.6.5",
    "tiktoken>=0.4.0",
    "jinja2>=3.1.2",
    "sseclient-py>=1.8.0",
    "pynvml>=11.5.0",
    "pandas>=2.0.3",
    "tabulate>=0.9.0",
]

requires-python = ">=3.10"
license = {text = "MIT"}

[tool.pdm.scripts]
start = "python3 src"
setup = "python3 scripts/setup.py"
clean = "python3 scripts/clean.py"
debug = "python3 scripts/debug.py"
dev = "python3 scripts/dev.py"
